requirements.txt
asyncpg
coc.py
disnake
requests
